refactor(core): organize DeepSeek model files
